minor Warlock reorg
