+++ Diagnostics/CommunityExpansionBlueprints/Assets.txt
@@ -4514,7 +4514,7 @@ Turn_12C~Sewers~MOD	RoomBlueprint	RoomBlueprint	4f5ba24c-2ccc-5b3b-820c-6efa1002
 Turn_12C~Stronghold~MOD	RoomBlueprint	RoomBlueprint	a66260e4-487d-5ca5-ac89-25c882c8ad75
 Turn_12C~ValleyPalace~MOD	RoomBlueprint	RoomBlueprint	09a5b4b5-749e-51ca-b59b-44f8e0b13be4
 AcidClaws	SpellDefinition	SpellDefinition	016894e0-1c5d-5fe5-a320-e99a29f0c035
-AHPactMarkSpellBuilder	SpellDefinition	SpellDefinition	ef9a57d2-4f91-5129-830f-44a15a4e2697
+AHPactMarkSpell	SpellDefinition	SpellDefinition	e59222d7-222b-571c-86df-47fff2c85cd9
 AirBlast	SpellDefinition	SpellDefinition	021bc2a7-b978-54b0-9d5a-1b9e50123cb5
 AlchemistBoldnessElixir	SpellDefinition	SpellDefinition	16a3f274-9fb0-5bbd-a673-281c2c845a5a
 AlchemistFlyElixir	SpellDefinition	SpellDefinition	d5484ca8-7cb5-5ef8-9c68-024346d83f26
