+++ Diagnostics/CommunityExpansionBlueprints/SpellDefinition/AHPactMarkSpell.json
@@ -347,7 +347,7 @@
     "usedInValleyDLC": false
   },
   "contentCopyright": "UserContent",
-  "guid": "ef9a57d2-4f91-5129-830f-44a15a4e2697",
+  "guid": "e59222d7-222b-571c-86df-47fff2c85cd9",
   "contentPack": 9999,
-  "name": "AHPactMarkSpellBuilder"
+  "name": "AHPactMarkSpell"
 }
