+++ Diagnostics/CommunityExpansionBlueprints/SpellListDefinition/ClassWarlockSpellList.json
@@ -19,7 +19,7 @@
       "$type": "SpellListDefinition+SpellsByLevelDuplet, Assembly-CSharp",
       "level": 1,
       "spells": [
-        "Definition:AHPactMarkSpellBuilder:ef9a57d2-4f91-5129-830f-44a15a4e2697",
+        "Definition:AHPactMarkSpell:e59222d7-222b-571c-86df-47fff2c85cd9",
         "Definition:CharmPerson:9c98e5cba0f8b1649b3217a97cbc111b",
         "Definition:ComprehendLanguages:78c2b416e515641468aa1c384db23535",
         "Definition:ExpeditiousRetreat:3713ff85f9f53f2459ae8d7a976a24b8",
